% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalEllipse.R
\name{ellipsoidalNEllipse}
\alias{ellipsoidalNEllipse}
\title{Ellipsoidal N-Ellipse}
\usage{
ellipsoidalNEllipse(
  ps,
  d,
  n.pts = 50,
  a = 6378137,
  f = 1/298.257223563,
  n.iters = 2 + max(6, ceiling(log2(d)))
)
}
\arguments{
\item{ps}{a two column matrix giving the longitude and latitude
of the focal points.}

\item{d}{the combined distance from the two focal points to a
point on the ellipse.}

\item{n.pts}{minimum number of points to generate.}

\item{a}{major (equatorial) radius of the ellipsoid (default is
for WGS84).}

\item{f}{ellipsoid flattening (default is for WGS84).}

\item{n.iters}{number of bisection iterations.}
}
\value{
an n x 2 matrix of longitudes and latitudes of points on
the ellipse.
}
\description{
Calculate points on an n-ellipse on the surface of the Earth.
}
\details{
An ellipsoidal n-ellipse is the locus of points on an ellipsoid
for which the sum of geodesic distances to n focal points on the
ellipsoid is a constant.

Given the latitude and longitude of n focal points \code{ps}, this
function calculates the longitude and latitude of a sequence of
points of the Earth for which the sum of the geodesic distances to
the focal points is \code{d}.

The final point is identical to the first.

This implementation constructs points on the n-ellipse by
computing the distance to the n-ellipse along n evenly spaced
bearings from the centroid of the foci using a bisection
procedure. The accuracy of the procedure is approximately
d/2^n.iters. The centroid is calculated assuming the Earth is
spherical and so this function may fail for small \code{d}.

This function is based upon \code{\link[geosphere:distGeo]{geosphere::distGeo()}} and an
explanation of the parameters \code{a} and \code{f} can be found there.
}
\examples{
p1 <- c(140,-45)
p2 <- c(143,-40)
p3 <- c(138,-43)
plot(rbind(p1,p2,p3),
     xlab="Lon",ylab="Lat",xlim=c(135,145),ylim=c(-47,-38),
     pch=16,col="red")
points(t(sphericalCentroid(rbind(p1,p2,p3))),pch=16,col="blue")
ps <- ellipsoidalNEllipse(rbind(p1,p2,p3),1400000)
points(ps,type="l")
ps <- ellipsoidalNEllipse(rbind(p1,p2,p3),1000000)
lines(ps,type="l")
ps <- ellipsoidalNEllipse(rbind(p1,p2,p3),800000)
lines(ps,type="l")
}
\seealso{
\code{\link[=ellipsoidalEllipse]{ellipsoidalEllipse()}}
}
