% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalEllipse.R
\name{focalIntersection}
\alias{focalIntersection}
\title{Focal Intersection}
\usage{
focalIntersection(p1, p2, d1, d2)
}
\arguments{
\item{p1}{a vector of length 2 giving the longitude and latitude
of the first focal point}

\item{p2}{a vector of length 2 giving the longitude and latitude
of the second focal point}

\item{d1}{a vector of great circle distances to the first focal
point.}

\item{d2}{a vector of great circle distances to the second focal
point.}
}
\value{
a list of two, two-column matrices representing the
longitude and latitude of the two alternate points.
}
\description{
Find the points on a unit sphere that are specified great circle
distances from two focal points.
}
\details{
Given two focal points \code{p1} and \code{p2} on the surface of the unit
sphere, this function calculates the two points that are distance
\code{d1} from \code{p1} and distance \code{d2} from \code{p2}. The vectors of
distances \code{d1} and \code{d2} must be the same length, and two alternate
points are given for each pairs of great circle distances.
}
