% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalEllipse.R
\name{sphericalCentroid}
\alias{sphericalCentroid}
\title{Spherical Centroid}
\usage{
sphericalCentroid(ps, n.iters = 20)
}
\arguments{
\item{ps}{a two column matrix giving the longitude and latitude of
the points.}

\item{n.iters}{number of iterations.}
}
\value{
A vector of length 2 giving the longitude and latitude of
the centroid.
}
\description{
Find the centroid of n points on a sphere.
}
\details{
Calculate the centroid of n points on a sphere by an iterative
method.
}
