% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalEllipse-package.R
\docType{package}
\name{SphericalEllipse-package}
\alias{SphericalEllipse}
\alias{SphericalEllipse-package}
\title{SphericalEllipse: Construct a spherical ellipse on the surface of the Earth}
\description{
Calculate the longitude and latitude of a set of points lying on a spherical ellipse on the lying surface of the Earth, assuming the Earth is a sphere.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/SphericalEllipse}
  \item Report bugs at \url{https://github.com/SWotherspoon/SphericalEllipse/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
