% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SphericalEllipse-package}
\alias{SphericalEllipse-package}
\title{Spherical Ellipse}
\description{
Calculate the longitude and latitude of a set of points lying on a
spherical ellipse lying on the surface of the Earth, assuming the
Earth is a sphere.
}
\author{
S. Wotherspoon
}
